third_party/WebKit/LayoutTests/http/tests/devtools/network/json-preview-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests that resources with JSON MIME types are previewed with the JSON viewer.


Running: test1
Creating a NetworkRequest with type: application/json
Its previewer is searchable: true
Its previewer is the JSON previewer: true

Running: test2
Creating a NetworkRequest with type: application/vnd.document+json
Its previewer is searchable: true
Its previewer is the JSON previewer: true
